requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.3.0
chromadb==0.4.18
requests==2.31.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
firebase-admin==6.2.0
signal-protocol==0.2.2
# lemon-squeezy-python==0.1.0  # Placeholder, replace with correct package
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
pytest==7.4.3
httpx
python-multipart==0.0.6
fastapi-limiter==0.1.6
aiofiles==23.2.1